fixed some things

+++ ft_strsplit.c
@@ -6,7 +6,7 @@
 /*   By: chermist <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/03 11:32:02 by chermist          #+#    #+#             */
-/*   Updated: 2019/01/11 22:01:04 by chermist         ###   ########.fr       */
+/*   Updated: 2019/05/27 21:55:58 by chermist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 char	**ft_strsplit(char const *s, char c)
 {
 	char	**str;
-	int		words;
-	int		j;
-	int		k;
+	size_t	words;
+	size_t	j;
+	size_t	k;
 
 	if (!s || !(str = (char**)malloc(sizeof(*str) *
 			((words = ft_wcount(s, c)) + 1))))
