+++ ft_wcount.c
@@ -6,7 +6,7 @@
 /*   By: chermist <chermist@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/04 19:28:56 by chermist          #+#    #+#             */
-/*   Updated: 2018/12/13 01:17:37 by chermist         ###   ########.fr       */
+/*   Updated: 2019/05/27 20:23:23 by chermist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int		ft_wcount(char const *s, char c)
 {
 	int		countw;
-	int		i;
+	size_t		i;
 
 	i = 0;
 	countw = 0;
